+++ .gitignore
@@ -1,3 +1,5 @@
 venv
 .vscode
 __pycache__/
+.pytest_cache
+.coverage
